requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
loguru~=0.4.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyee==8.1.0
pyparsing==3.0.7
pytest==7.1.1
pytest-assume==2.4.3
pytest-base-url==2.0.0
pytest-ordering==0.6
python-slugify==6.1.1
PyYAML~=6.0
requests~=2.27.0
six==1.16.0
text-unidecode==1.3
tomli==2.0.1
urllib3==1.26.9
websockets==10.1
win32-setctime==1.1.0
mitmproxy~=8.0.0
websocket-client~=1.3.2
jsonpath_ng~=1.5.3
nested_lookup~=0.2.23
selenium~=4.2.0
DingtalkChatbot~=1.5.3
PyMySQL~=1.0.2
cx_Oracle~=8.3.0
PyHive~=0.6.5
playwright==1.21.0
pytest-playwright==0.3.0
jmespath~=0.9.5
python-dateutil~=2.8.2
deepwisdom==0.2.4
